+++ app/src/main/java/eu/faircode/email/DB.java
@@ -71,7 +71,7 @@ import io.requery.android.database.sqlite.SQLiteDatabase;
 // https://developer.android.com/topic/libraries/architecture/room.html
 
 @Database(
-        version = 239,
+        version = 240,
         entities = {
                 EntityIdentity.class,
                 EntityAccount.class,
@@ -2404,6 +2404,12 @@ public abstract class DB extends RoomDatabase {
                             Log.e(ex);
                         }
                     }
+                }).addMigrations(new Migration(239, 240) {
+                    @Override
+                    public void migrate(@NonNull SupportSQLiteDatabase db) {
+                        logMigration(startVersion, endVersion);
+                        db.execSQL("ALTER TABLE `search` ADD COLUMN `order` INTEGER");
+                    }
                 }).addMigrations(new Migration(998, 999) {
                     @Override
                     public void migrate(@NonNull SupportSQLiteDatabase db) {
@@ -2427,7 +2433,7 @@ public abstract class DB extends RoomDatabase {
                     long start = new Date().getTime();
                     StringBuilder sb = new StringBuilder();
                     SupportSQLiteDatabase sdb = db.getOpenHelper().getWritableDatabase();
-                    String mode = (BuildConfig.DEBUG || !BuildConfig.PLAY_STORE_RELEASE ? "RESTART" : "PASSIVE");
+                    String mode = (true ? "RESTART" : "PASSIVE");
                     try (Cursor cursor = sdb.query("PRAGMA wal_checkpoint(" + mode + ");")) {
                         if (cursor.moveToNext()) {
                             for (int i = 0; i < cursor.getColumnCount(); i++) {
@@ -2439,8 +2445,8 @@ public abstract class DB extends RoomDatabase {
                     }
 
                     long elapse = new Date().getTime() - start;
-                    EntityLog.log(context, EntityLog.Type.Debug,
-                            "PRAGMA wal_checkpoint(" + mode + ")=" + sb + " elapse=" + elapse + " ms");
+                    EntityLog.log(context, "PRAGMA wal_checkpoint(" + mode + ")=" + sb +
+                            " elapse=" + elapse + " ms");
                 } catch (Throwable ex) {
                     Log.e(ex);
                 }