+++ app/src/main/java/eu/faircode/email/GmailState.java
@@ -39,7 +39,7 @@ public class GmailState {
     private long acquired;
 
     static final String TYPE_GOOGLE = "com.google";
-    private static final long TOKEN_LIFETIME = 45 * 60 * 1000L; // milliseconds
+    private static final long TOKEN_LIFETIME = 60 * 60 * 1000L; // milliseconds
 
     private GmailState(String token, long acquired) {
         this.token = token;
@@ -60,20 +60,27 @@ public class GmailState {
             return acquired + TOKEN_LIFETIME;
     }
 
-    void refresh(@NonNull Context context, @NonNull String user, boolean expire, long keep_alive)
+    void refresh(@NonNull Context context, String id, @NonNull String user, boolean forceRefresh)
             throws AuthenticatorException, OperationCanceledException, IOException {
+        long now = new Date().getTime();
         Long expiration = getAccessTokenExpirationTime();
-        if (expiration != null && expiration - keep_alive < new Date().getTime()) {
-            EntityLog.log(context, "Force invalidation" +
-                    " expiration=" + new Date(expiration) +
-                    " keep-alive=" + (keep_alive / 60 / 1000) + "m");
-            expire = true;
-        }
+        boolean needsRefresh = (expiration != null && expiration < now);
 
-        if (expire)
+        if (!needsRefresh && forceRefresh &&
+                expiration != null &&
+                expiration - ServiceAuthenticator.MIN_FORCE_REFRESH_INTERVAL < now)
+            needsRefresh = true;
+
+
+        EntityLog.log(context, EntityLog.Type.Debug, "Token user=" + id + ":" + user +
+                " expiration=" + (expiration == null ? null : new Date(expiration)) +
+                " need=" + needsRefresh +
+                " force=" + forceRefresh);
+
+        if (needsRefresh)
             try {
                 if (token != null) {
-                    EntityLog.log(context, "Invalidating token user=" + user);
+                    EntityLog.log(context, "Invalidating token user=" + id + ":" + user);
                     AccountManager am = AccountManager.get(context);
                     am.invalidateAuthToken(TYPE_GOOGLE, token);
                 }
@@ -85,9 +92,9 @@ public class GmailState {
 
         Account account = getAccount(context, user.replace("recent:", ""));
         if (account == null)
-            throw new AuthenticatorException("Account not found for " + user);
+            throw new AuthenticatorException("Account not found for " + id + ":" + user);
 
-        EntityLog.log(context, "Getting token user=" + user);
+        EntityLog.log(context, "Getting token user=" + id + ":" + user);
         AccountManager am = AccountManager.get(context);
         String newToken = am.blockingGetAuthToken(
                 account,
@@ -100,7 +107,7 @@ public class GmailState {
         }
 
         if (token == null)
-            throw new AuthenticatorException("No token for " + user);
+            throw new AuthenticatorException("No token for " + id + ":" + user);
     }
 
     static Account getAccount(Context context, String user) {