+++ app/src/main/java/eu/faircode/email/ServiceAuthenticator.java
@@ -47,7 +47,6 @@ public class ServiceAuthenticator extends Authenticator {
     private Context context;
     private int auth;
     private String provider;
-    private long keep_alive;
     private String user;
     private String password;
     private IAuthenticated intf;
@@ -56,15 +55,16 @@ public class ServiceAuthenticator extends Authenticator {
     static final int AUTH_TYPE_GMAIL = 2;
     static final int AUTH_TYPE_OAUTH = 3;
 
+    static final long MIN_FORCE_REFRESH_INTERVAL = 15 * 60 * 1000L;
+
     ServiceAuthenticator(
             Context context,
-            int auth, String provider, int keep_alive,
+            int auth, String provider,
             String user, String password,
             IAuthenticated intf) {
         this.context = context.getApplicationContext();
         this.auth = auth;
         this.provider = provider;
-        this.keep_alive = keep_alive * 60 * 1000L;
         this.user = user;
         this.password = password;
         this.intf = intf;
@@ -82,14 +82,16 @@ public class ServiceAuthenticator extends Authenticator {
                 Log.e(ex);
         }
 
-        Log.i(user + " returning " + (auth == AUTH_TYPE_PASSWORD ? "password" : "token"));
+        Log.i(user + " returning " +
+                (auth == AUTH_TYPE_PASSWORD ? "password" : "token") +
+                (BuildConfig.DEBUG ? "=" + token : ""));
         return new PasswordAuthentication(user, token);
     }
 
-    String refreshToken(boolean expire) throws AuthenticatorException, OperationCanceledException, IOException, JSONException, MessagingException {
+    String refreshToken(boolean forceRefresh) throws AuthenticatorException, OperationCanceledException, IOException, JSONException, MessagingException {
         if (auth == AUTH_TYPE_GMAIL) {
             GmailState authState = GmailState.jsonDeserialize(password);
-            authState.refresh(context, user, expire, keep_alive);
+            authState.refresh(context, "android", user, forceRefresh);
             Long expiration = authState.getAccessTokenExpirationTime();
             if (expiration != null)
                 EntityLog.log(context, user + " token expiration=" + new Date(expiration));
@@ -104,7 +106,7 @@ public class ServiceAuthenticator extends Authenticator {
             return authState.getAccessToken();
         } else if (auth == AUTH_TYPE_OAUTH && provider != null) {
             AuthState authState = AuthState.jsonDeserialize(password);
-            OAuthRefresh(context, provider, authState, expire, keep_alive);
+            OAuthRefresh(context, provider, user, authState, forceRefresh);
             Long expiration = authState.getAccessTokenExpirationTime();
             if (expiration != null)
                 EntityLog.log(context, user + " token expiration=" + new Date(expiration));
@@ -121,43 +123,43 @@ public class ServiceAuthenticator extends Authenticator {
             return password;
     }
 
-    void checkToken() {
-        Long expiration = null;
-
+    Long getAccessTokenExpirationTime() {
         try {
             if (auth == AUTH_TYPE_GMAIL) {
                 GmailState authState = GmailState.jsonDeserialize(password);
-                expiration = authState.getAccessTokenExpirationTime();
+                return authState.getAccessTokenExpirationTime();
             } else if (auth == AUTH_TYPE_OAUTH) {
                 AuthState authState = AuthState.jsonDeserialize(password);
-                expiration = authState.getAccessTokenExpirationTime();
+                return authState.getAccessTokenExpirationTime();
             }
         } catch (JSONException ex) {
             Log.e(ex);
         }
-
-        long slack = Math.min(keep_alive, 30 * 60 * 1000L);
-        if (expiration != null && expiration - slack < new Date().getTime())
-            throw new IllegalStateException(Log.TOKEN_REFRESH_REQUIRED);
+        return null;
     }
 
     interface IAuthenticated {
         void onPasswordChanged(Context context, String newPassword);
     }
 
-    private static void OAuthRefresh(Context context, String id, AuthState authState, boolean expire, long keep_alive)
+    private static void OAuthRefresh(Context context, String id, String user, AuthState authState, boolean forceRefresh)
             throws MessagingException {
         try {
+            long now = new Date().getTime();
             Long expiration = authState.getAccessTokenExpirationTime();
-            if (expiration != null && expiration - keep_alive < new Date().getTime()) {
-                EntityLog.log(context, "OAuth force refresh" +
-                        " expiration=" + new Date(expiration) +
-                        " keep_alive=" + (keep_alive / 60 / 1000) + "m");
+            boolean needsRefresh = (expiration != null && expiration < now);
+            if (needsRefresh)
                 authState.setNeedsTokenRefresh(true);
-            }
 
-            if (expire)
+            if (!needsRefresh && forceRefresh &&
+                    expiration != null &&
+                    expiration - ServiceAuthenticator.MIN_FORCE_REFRESH_INTERVAL < now)
                 authState.setNeedsTokenRefresh(true);
+
+            EntityLog.log(context, EntityLog.Type.Debug, "Token user=" + id + ":" + user +
+                    " expiration=" + (expiration == null ? null : new Date(expiration)) +
+                    " need=" + needsRefresh + "/" + authState.getNeedsTokenRefresh() +
+                    " force=" + forceRefresh);
 
             ClientAuthentication clientAuth;
             EmailProvider provider = EmailProvider.getProvider(context, id);
@@ -169,7 +171,7 @@ public class ServiceAuthenticator extends Authenticator {
             ErrorHolder holder = new ErrorHolder();
             Semaphore semaphore = new Semaphore(0);
 
-            Log.i("OAuth refresh id=" + id);
+            Log.i("OAuth refresh user=" + id + ":" + user);
             AuthorizationService authService = new AuthorizationService(context);
             authState.performActionWithFreshTokens(
                     authService,
@@ -184,12 +186,12 @@ public class ServiceAuthenticator extends Authenticator {
                     });
 
             semaphore.acquire();
-            Log.i("OAuth refreshed id=" + id);
+            Log.i("OAuth refreshed user=" + id + ":" + user);
 
             if (holder.error != null)
                 throw holder.error;
         } catch (Exception ex) {
-            throw new MessagingException("OAuth refresh id=" + id, ex);
+            throw new MessagingException("OAuth refresh id=" + id + ":" + user, ex);
         }
     }